% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastqcShinyApp.R
\name{fastqcShiny}
\alias{fastqcShiny}
\title{Run fastQC shiny app}
\usage{
fastqcShiny(fastqcInput = NULL)
}
\arguments{
\item{fastqcInput}{can be a \code{FastqcFileList},
 \code{fastqcDataList},
or simply a \code{character} vector
 of paths to fastqc files.}
}
\value{
UI data for fastQC shiny.
}
\description{
Returns a shiny app
interface to parse many fastQC
objects using the package ngsReports
}
\details{
Currently some plots can
take a while to render if the
\code{FastqcDataList} passed to
\code{fastqcInput} has many elements
}
\examples{
\dontrun{
Get the files included with the package ngsReports
packageDir <- system.file("extdata", package = "ngsReports")
fileList <- list.files(packageDir, pattern = "fastqc.zip", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- getFastqcData(fileList)

# Run the Shiny app
fastqcShiny(fdl)
}

}
